backend/pyproject.toml
[project]
name = "ai-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.11.4",
    "fastapi>=0.135.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.20",
    "langchain-litellm>=0.6.1",
    "langgraph>=1.1.3",
    "litellm==1.82.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
